+++ web/connection/safe_conn.go
@@ -165,15 +165,3 @@ func (sc *SafeConn) ReadJSON(v interface{}) error {
 func (sc *SafeConn) SetReadDeadline(t time.Time) error {
 	return sc.conn.SetReadDeadline(t)
 }
-
-func (sc *SafeConn) GetConn() *websocket.Conn {
-	sc.mu.Lock()
-	defer sc.mu.Unlock()
-	return sc.conn
-}
-
-// SetCloseHandler forwards to the underlying connection (used by the
-// terminal sessions to clean up on close).
-func (sc *SafeConn) SetCloseHandler(h func(code int, text string) error) {
-	sc.conn.SetCloseHandler(h)
-}